requirements.txt
asn1crypto==0.24.0
#astroid==2.3.1
autopep8==1.4.4
certifi==2019.9.11
cffi==1.12.3
cryptography==2.7
dnspython==1.16.0
emoji==0.5.2
future==0.17.1
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pep8==1.7.1
#pkg-resources==0.0.0
pycodestyle==2.5.0
pycparser==2.19
pylint
PySocks==1.7.1
python-telegram-bot==11.1.0
schedule==0.6.0
six==1.12.0
SQLAlchemy==1.3.8
typed-ast==1.4.0
urllib3==1.25.6
wrapt==1.11.2
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
WTForms==2.2.1